REFACTOR: seperated blockchain package to block.go, chain.go
