+++ blockchain/block.go
@@ -1,0 +1,27 @@
+package blockchain
+
+import (
+	"crypto/sha256"
+	"fmt"
+)
+
+type Block struct {
+	Data     string `json:"data"`
+	Hash     string `json:"hash"`
+	PrevHash string `json:"prevHash,omitempty"`
+	Height   int    `json:"height"`
+}
+
+func createBlock(data string) *Block {
+	block := Block{
+		Data:     data,
+		Hash:     "",
+		PrevHash: BlockChain().NewestHash,
+		Height:   BlockChain().Height,
+	}
+
+	payload := block.Data + block.PrevHash + fmt.Sprint(block.Height)
+	block.Hash = fmt.Sprintf("%x", sha256.Sum256([]byte(payload)))
+	
+	return &block
+}
