--- blockchain/blockchain.go
@@ -1,66 +1,0 @@
-package blockchain
-
-import (
-	"crypto/sha256"
-	"fmt"
-	"log"
-	"sync"
-)
-
-type Block struct {
-	Data string `json:"data"`
-	Hash string `json:"hash"`
-	PrevHash string `json:"prevHash,omitempty"`
-	Height int `json:"height"`
-}
-
-type blockchain struct {
-	blocks []*Block
-}
-
-var b *blockchain
-var once sync.Once
-
-func (b *Block) calculateHash() {
-	hash := sha256.Sum256([]byte(b.Data + b.PrevHash))
-	b.Hash = fmt.Sprintf("%x", hash)
-}
-
-func getLastHash() string {
-	totalBlocks := len(GetBlockChain().blocks)
-	if totalBlocks == 0 {
-		return ""
-	}
-	return GetBlockChain().blocks[totalBlocks - 1].Hash
-}
-
-func createBlock(data string) *Block {
-	newBlock := Block{data, "", getLastHash(), len(GetBlockChain().blocks)+1}
-	newBlock.calculateHash()
-	return &newBlock
-}
-
-func (b *blockchain) AddBlock(data string) {
-	b.blocks = append(b.blocks, createBlock(data))
-}
-
-func GetBlockChain() *blockchain {
-	if b == nil {
-		once.Do(func() {
-			b = &blockchain{}
-			b.AddBlock("Genesis Block")
-		})
-	}
-	return b
-}
-
-func (b *blockchain) AllBlocks() []*Block {
-	return b.blocks
-}
-
-func (b *blockchain) GetBlock(height int) *Block {
-	if height > len(b.blocks) {
-		log.Panic("Index Out Of Range")
-	}
-	return b.blocks[height-1]
-}
