+++ blockchain/chain.go
@@ -1,0 +1,29 @@
+package blockchain
+
+import (
+	"sync"
+)
+
+type blockchain struct {
+	NewestHash string `json: "newestHash"`
+	Height     int    `json:"height"`
+}
+
+var b *blockchain
+var once sync.Once
+
+func (b *blockchain) AddBlock(data string) {
+	block := createBlock(data)
+	b.NewestHash = block.Hash
+	b.Height = block.Height
+}
+
+func BlockChain() *blockchain {
+	if b == nil {
+		once.Do(func() {
+			b = &blockchain{"", 0}
+			b.AddBlock("Genesis Block")
+		})
+	}
+	return b
+}
